Extract partial LukionOppiaineEditor components
